refactor(auth): move auth into features/auth with a login/register split
